Proyecto/requirements.txt
gradio
openai_Whisper
traslate
python-dotenv
elevenlabs
langdetect